added shop screen and players move, added enemies (dragon, spikes, hunting spiter with GDX AI)

+++ core/src/com/ledzinygamedevelopment/fallingman/screens/windows/DefaultWindow.java
@@ -24,8 +24,8 @@ public class DefaultWindow extends Sprite {
     private float height;
     private float timer;
     private boolean drawText;
-    private int gold;
-    private int wholeDistance;
+    private long gold;
+    private long wholeDistance;
     private boolean tapGrow;
     private float scale;
     private boolean tapExist;
